+++ lib/Utility/PodcastService.php
@@ -7,7 +7,7 @@
  * later. See the COPYING file.
  *
  * @author Pauli Järvinen <pauli.jarvinen@gmail.com>
- * @copyright Pauli Järvinen 2021 - 2024
+ * @copyright Pauli Järvinen 2021 - 2025
  */
 
 namespace OCA\Music\Utility;
@@ -19,7 +19,8 @@ use OCA\Music\BusinessLayer\PodcastEpisodeBusinessLayer;
 use OCA\Music\Db\PodcastChannel;
 use OCA\Music\Db\PodcastEpisode;
 use OCA\Music\Db\SortBy;
-
+use OCP\Files\File;
+use OCP\Files\Folder;
 
 class PodcastService {
 	private PodcastChannelBusinessLayer $channelBusinessLayer;
@@ -260,5 +261,145 @@ class PodcastService {
 		}
 		// return the episodes in inverted chronological order (newest first)
 		return \array_reverse($episodes);
+	}
+
+	/**
+	 * export all the podcast channels of a user to an OPML file
+	 * @param string $userId user
+	 * @param Folder $userFolder home dir of the user
+	 * @param string $folderPath target parent folder path
+	 * @param string $filename target file name
+	 * @param string $collisionMode action to take on file name collision,
+	 *								supported values:
+	 *								- 'overwrite' The existing file will be overwritten
+	 *								- 'keepboth' The new file is named with a suffix to make it unique
+	 *								- 'abort' (default) The operation will fail
+	 * @return string path of the written file
+	 * @throws \OCP\Files\NotFoundException if the $folderPath is not a valid folder
+	 * @throws \RuntimeException on name conflict if $collisionMode == 'abort'
+	 * @throws \OCP\Files\NotPermittedException if the user is not allowed to write to the given folder
+	 */
+	public function exportToFile(
+		string $userId, Folder $userFolder, string $folderPath, string $filename, string $collisionMode='abort') : string {
+		$targetFolder = FilesUtil::getFolderFromRelativePath($userFolder, $folderPath);
+
+		$filename = FilesUtil::sanitizeFileName($filename, ['opml']);
+
+		$file = FilesUtil::createFile($targetFolder, $filename, $collisionMode);
+
+		$channels = $this->channelBusinessLayer->findAll($userId, SortBy::Name);
+
+		$content = self::channelsToOpml($channels);
+		$file->putContent($content);
+
+		return $userFolder->getRelativePath($file->getPath());
+	}
+
+	/**
+	 * @param PodcastChannel[] $channels
+	 */
+	private static function channelsToOpml(array $channels) : string {
+		$dom = new \DOMDocument('1.0', 'UTF-8');
+		$dom->formatOutput = true;
+
+		$rootElem = $dom->createElement('opml');
+		$rootElem->setAttribute('version', '1.0');
+		$dom->appendChild($rootElem);
+
+		$headElem = $dom->createElement('head');
+		$titleElem = $dom->createElement('title', 'Podcast channels from ownCloud/Nextcloud Music');
+		$now = new \DateTime();
+		$dateCreatedElem = $dom->createElement('dateCreated', $now->format(\DateTime::RFC822));
+		$headElem->appendChild($titleElem);
+		$headElem->appendChild($dateCreatedElem);
+		$rootElem->appendChild($headElem);
+
+		$bodyElem = $dom->createElement('body');
+		foreach ($channels as $channel) {
+			$outlineElem = $dom->createElement('outline');
+			$outlineElem->setAttribute('type', 'rss');
+			$outlineElem->setAttribute('text', $channel->getTitle());
+			$outlineElem->setAttribute('title', $channel->getTitle());
+			$outlineElem->setAttribute('xmlUrl', $channel->getRssUrl());
+			$outlineElem->setAttribute('htmlUrl', $channel->getLinkUrl());
+			$bodyElem->appendChild($outlineElem);
+		}
+		$rootElem->appendChild($bodyElem);
+
+		return $dom->saveXML();
+	}
+
+	/**
+	 * import podcast channels from an OPML file
+	 * @param string $userId user
+	 * @param Folder $userFolder user home dir
+	 * @param string $filePath path of the file to import
+	 * @return array with three keys:
+	 * 			- 'channels': Array of PodcastChannel objects imported from the file
+	 * 			- 'not_changed_count': An integer showing the number of channels in the file which were already subscribed by the user
+	 * 			- 'failed_count': An integer showing the number of entries in the file which were not valid URLs
+	 * @throws \OCP\Files\NotFoundException if the $filePath is not a valid file
+	 * @throws \UnexpectedValueException if the $filePath points to a file of unsupported type
+	 */
+	public function importFromFile(string $userId, Folder $userFolder, string $filePath, ?callable $progressCallback = null) : array {
+		$channelUrls = self::parseOpml($userFolder, $filePath);
+
+		$channels = [];
+		$existingCount = 0;
+		$failedCount = 0;
+
+		foreach ($channelUrls as $rssUrl) {
+			$channelResult = $this->subscribe($rssUrl, $userId);
+			if (!empty($channelResult['channel'])) {
+				$channels[] = $channelResult['channel'];
+			} else if ($channelResult['status'] == self::STATUS_ALREADY_EXISTS) {
+				$existingCount++;
+			} else {
+				$failedCount++;
+			}
+
+			if ($progressCallback !== null) {
+				$channelResult['rss'] = $rssUrl;
+				$progressCallback($channelResult);
+			}
+		}
+
+		return [
+			'channels' => $channels,
+			'not_changed_count' => $existingCount,
+			'failed_count' => $failedCount
+		];
+	}
+
+	/**
+	 * @return string[] RSS URLs
+	 */
+	public static function parseOpml(Folder $userFolder, string $filePath) : array {
+		$rssUrls = [];
+
+		$file = self::getFile($userFolder, $filePath);
+		$rootNode = \simplexml_load_string($file->getContent(), \SimpleXMLElement::class, LIBXML_NOCDATA);
+		if ($rootNode === false) {
+			throw new \UnexpectedValueException('the file is not in valid OPML format');
+		}
+
+		$rssNodes = $rootNode->xpath("/opml/body//outline[@type='rss']");
+
+		foreach ($rssNodes as $node) {
+			$rssUrls[] = (string)$node->attributes()['xmlUrl'];
+		}
+
+		return $rssUrls;
+	}
+
+	/**
+	 * @throws \OCP\Files\NotFoundException if the $path does not point to a file under the $baseFolder
+	 */
+	private static function getFile(Folder $baseFolder, string $path) : File {
+		$node = $baseFolder->get($path);
+		if (!($node instanceof File)) {
+			throw new \OCP\Files\NotFoundException();
+		}
+		return $node;
 	}
 }