+++ templates/partials/navigationitem.php
@@ -96,6 +96,12 @@
 					<li ng-click="$parent.addPodcast()">
 						<a><span class="icon-add icon"></span><span translate>Add from RSS feed</span></a>
 					</li>
+					<li ng-click="$parent.importPodcastsFromFile()">
+						<a><span class="icon-from-file icon svg"></span><span translate>Import from file</span></a>
+					</li>
+					<li ng-click="$parent.exportPodcastsToFile($event)" ng-class="{ disabled: !$parent.anyPodcastChannels() }">
+						<a><span class="icon-to-file icon svg"></span><span translate>Export to file</span></a>
+					</li>
 					<li ng-click="$parent.reloadPodcasts($event)" ng-class="{ disabled: !$parent.anyPodcastChannels() }">
 						<a><span class="icon-reload icon svg"></span><span translate>Reload channels</span></a>
 					</li>